exercise 4.21 done

+++ part4/bloglist/controllers/blogs.js
@@ -33,8 +33,20 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blogBelongsToLoggedUser =
+    user.blogs.findIndex((blog) => blog.toString() == request.params.id) > -1
+  if (blogBelongsToLoggedUser) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: "unauthorized" })
+  }
 })
 
 blogsRouter.put("/:id", async (request, response) => {
